Clear up most swiftlint warnings
